Refactor Greenplum specific testcode to a new file
regress.c is an upstream file, and all Greenplum additions can
cause conflicts as we merge with PostgreSQL. This refactors all
GPDB specific code into a new file, regress_gp.c, to keep the
upstream file as close to upstream as possible (with backports).
The new file gets compiled and loaded just like regress.c, so
no change in how it works.

Also remove an unused function, perform rudimentary codereview
on the Greenplum tests and massage regress.c slightly to make
it closer to upstream.

(cherry picked from commit 01304a76d1717debec55972503b62d88dcd0db04)
